requirements.txt
# 核心框架
streamlit>=1.28.0

# 大语言模型
transformers>=4.35.0
torch>=2.0.0

# 网页处理
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# OCR功能
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 其他工具
python-dotenv>=1.0.0